+++ utmosv2/_settings/__init__.py
@@ -1,4 +1,4 @@
-from utmosv2.settings.config import (
+from utmosv2._settings._config import (
     configure_args,
     configure_defaults,
     configure_inference_args,
